+++ src/core.rs
@@ -1,0 +1,1 @@
+// TODO: structs like `Message`, `Reading` and so on.