+++ src/supervisor.rs
@@ -1,6 +1,6 @@
 //! Allows to monitor thread status and automatically respawn a crashed thread.
 
-use crate::message::{Message, Type};
+use crate::message::*;
 use crate::Result;
 use crossbeam_channel::Sender;
 use log::{error, info};
@@ -19,12 +19,12 @@ where
     thread::Builder::new().name(name.clone()).spawn(move || loop {
         // TODO: update thread status.
         info!("Running {}", &name);
-        tx.send(Message::now(Type::ReadLogged, &sensor, true)).unwrap();
+        tx.send(Composer::new(&sensor).value(true).into()).unwrap();
         match f() {
             Ok(_) => error!("Thread {} has finished unexpectedly", &name),
             Err(error) => error!("Thread {} crashed: {:?}", &name, error),
         }
-        tx.send(Message::now(Type::ReadLogged, &sensor, false)).unwrap();
+        tx.send(Composer::new(&sensor).value(false).into()).unwrap();
 
         // FIXME: https://github.com/eigenein/my-iot-rs/issues/47
         thread::sleep(Duration::from_secs(60));