+++ src/templates/reading.rs
@@ -1,10 +1,11 @@
 //! Renders single reading on the sensors page.
 
+use crate::message::*;
 use crate::templates::DATE_FORMAT;
 
 // TODO: title should be human-readable.
 markup::define! {
-    Reading<'a>(reading: &'a crate::message::Reading) {
+    Reading<'a>(reading: &'a Message) {
         div."column"."is-one-quarter" {
             a[href = {format!("/sensors/{}", &reading.sensor)} ] {
                 div.notification.reading.{reading.value.class()} {