+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This release brings the different database schema, which should work faster and 
 - Break: optimise database schema #50
 - Break: rename message types and introduce `Type::ReadSnapshot` #49
 - Break: change sensor value serialization #50
+- Break: welcome message `Composer` #50
 
 ## `0.28.0`